% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeTrainTest.R
\name{MakeTrainTest}
\alias{MakeTrainTest}
\title{Data-preprocessing and provide useful data for fitting}
\usage{
MakeTrainTest(
  outcome_name,
  features_index,
  data_train,
  data_test,
  y_method = FALSE,
  boxcox = FALSE
)
}
\arguments{
\item{outcome_name}{The outcome's name (Strings)}

\item{features_index}{The index of columns you use as features}

\item{data_train}{Training data}

\item{data_test}{Test data}

\item{y_method}{Decide how outcome will be standardized.
\code{y_method = "scale"}: scale the outcome by dividing with it's standard deviation (\code{y / sd(y)}).
\code{y_method = "center"}: centering the outcome (\code{y - mean(y)}).
\code{c("scale", "center")}: standardize the outcome (\code{(y - mean(y)) / sd(y)}).
If \code{FALSE}, nothing is applied to the outcome.}

\item{boxcox}{If \code{TRUE}, return transformed outcome
with Box-Cox transformation using \code{car} package.}
}
\value{
A list that has scaled data
and separate data in features and outcome as matrix.
\item{train}{Standardized training data}
\item{test}{Standardized test data}
\item{train_x}{Drop outcome from \code{train}}
\item{train_y}{outcome of training data
(if \code{y_method} was specified, It'll be a transformed value).}
\item{test_x}{Drop outcome from \code{test}}
\item{test_y}{Outcome of test data (not changed)}
\item{outcome_mean}{A value used for centering outcome
(mean of outcome in training data)}
\item{outcome_sd}{A value used for scaling outcome
(standard deviation of outcome in test data)}
}
\description{
This function can arrange the data (Move the outcome to the left),
standardize the training data and test data
using training data's information.
You can also choose whether you transform the outcome
with Box-Cox transformation or not.
If you transform the outcome, it also returns
mean value \code{outcome_mean}, standard deviation \code{outcome_sd}
and Box-Cox transformation parameter \code{lambda}
}
