% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reportWithin.R
\name{reportWithin}
\alias{reportWithin}
\title{Report descriptive statistics with within subjects' standard errors}
\usage{
reportWithin(data = NULL, wid = NULL, within = NULL, between = NULL,
  dv = NULL, conf.interval = 0.95)
}
\arguments{
\item{data}{A data frame in the long format}

\item{wid}{A string indicating the column name of the subjects (or observations)}

\item{within}{A vector of string of column names of the within variables
(repeated within the same observation)}

\item{between}{A vector of string of column names of the between variables (across observations)}

\item{dv}{A string indicating the column name of the dependent variable
(what is measured, recorded)}

\item{conf.interval}{The rate to use to compute the confident interval. Defaults to .95}
}
\value{
Return a data frame of the number of observation (N), mean, standard deviation (sd),
 corrected standard error (se), and confident interval (ci) by conditions.
}
\description{
A function cannibalized and slightly simplified from
\url{http://www.cookbook-r.com/Manipulating_data/Summarizing_data/}
}
\details{
As reportSE, the function compute simple descriptive data: number of observations (N), mean,
 standard deviation (sd), confident intervals (ci) and standard error (se).
 The main difference is that reportWithin compute the standard error to handle within subjects
 design. The function normalized the data to remove subjects variability and report intra-subject
 variability (see Loftus & Masson, 1994).

Within standard errors are computed according the method described by Cousineau (2005)
 and corrected with the method of Morey (2008).
}
\examples{
reportWithin()
}
\author{
Guillaume T. Vallet \email{gtvallet@gmail.com}, University of Montreal (Canada)
}
\references{
Loftus, G.R., & Masson, M.E.J. (1994). Using confidence intervals in within-subject designs. Psychonomic Bulletin & Review, 1(4), 476–490.

Cousineau, D. (2005). Confidence intervals in within‐subject designs:A simpler solution to Loftus and Masson’s method. Tutorials in Quantitative Methods for Psychology, 1, 42-45.

Morey, R. (2008). Confidence intervals from normalized data: A correction to Cousineau (2005). Tutorial in Quantitative Methods for Psychology, 4(2), 61–64.
}
\seealso{
\code{\link{reportSE}}
}
\keyword{summary}

