% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize}
\alias{standardize}
\title{Standardize training data and test data}
\usage{
standardize(test, train, y_method = c("center", "scale"))
}
\arguments{
\item{test}{Test data to standardize.}

\item{train}{Training data that is standardized
and used to standardize test data.}

\item{y_method}{Decide how outcome will be standardized.
If \code{FALSE}, nothing is applied.}
}
\value{
A list that has standardized training and test data,
which includes standard deviation and mean of outcome,
The outcome of test data is not changed.
\item{train}{Standardized training data}
\item{test}{Standardized test data}
\item{y_mean}{A value used for centering
(mean of outcome in training data)}
\item{y_sd}{A value used for scaling
(standard deviation of outcome in test data)}
}
\description{
Standardize training and test data by the information of training data.
Test data is standardized using the information of training data.
}
\details{
Standardization will be done by following formula:
\code{z = (x - mean(x)) / sd(x)}
}
\examples{
data_train <- data.frame(y = rnorm(10), x1 = rnorm(10), x2 = rnorm(10))
data_test <- data.frame(y = rnorm(10), x1 = rnorm(10), x2 = rnorm(10))
apply(data_train, 2, mean); apply(data_train, 2, sd)
data_scaled <- standardize(test = data_test,
train = data_train, y_method = "scale")
apply(data_scaled$train, 2, mean); apply(data_scaled$train, 2, sd)
}
