% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/strFrqTable.R
\name{strFrqTable}
\alias{strFrqTable}
\title{Format Frequency Tsables}
\usage{
strFrqTable(dt, target, vars = NULL, sep_symbol = "/")
}
\arguments{
\item{dt}{A data frame to process.}

\item{target}{A string of the column name to target
(i.e. the one used to compute the frequency table).}

\item{vars}{Optional, a string or vector of strings
of the column names of the variables to consider when
computing the frequency table.}

\item{sep_symbol}{Optional, the symbol to use in the
concatenation of the frequencies. Default to "/".}
}
\value{
Return a data frame with the frequencies concatenated
  in a string separated by the sep_symbol argument (default
  to '/').
}
\description{
Simple function to compute the frequency table of a given factor
 according to some variables (optional).
Return a data frame with the frequencies concatenate together.
It is of your responsability to identify (or rename) the levels
 of the target variable.
}
\examples{
# Create data frame faking a group of young and elderly adults.
data = rbind(
     data.frame(subject=1:10,
         age=round(rnorm(10, mean=23, sd=1.5), 0),
         education=round(rnorm(10, mean=15, sd=2.1), 0),
         sex=sample(1:2, size=10, replace=T, prob=c(.6,.4)),
         group='young'),
     data.frame(subject=11:20,
         age=round(rnorm(10, mean=71, sd=2), 0),
         education=round(rnorm(10, mean=14, sd=1.9), 0),
         sex=sample(1:2, size=10, replace=T, prob=c(.4,.6)),
         group='elderly'))
# Compute the frequency table for the gender variable.
strFrqTable(data, 'sex', sep_symbol="-")
# Compute the frequency table for the gender variable by group.
strFrqTable(data, target='sex', vars='group')
}
\author{
Guillaume T. Vallet \email{gtvallet@gmail.com}, University of Montreal (Canada)
}
\keyword{data,}
\keyword{descriptive,}
\keyword{frequency}
\keyword{table}

