% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/strMeanSd.R
\name{strMeanSd}
\alias{strMeanSd}
\title{Format means and standard deviations}
\usage{
strMeanSd(x, dmean = 2, dsd = 1)
}
\arguments{
\item{x}{A vector of numbers.}

\item{dmean}{Optional, the number of digits to use in rounding the mean.
Default to 2.}

\item{dsd}{Optional, the number of digits to use in rounding the
standard deviation. Default to 1.}
}
\value{
Return a string formatted as mean (standard deviation).
}
\description{
Very simple function to compute the mean and the standard
 deviation of a vector of numbers and format them as
 mean (standard deviation). The number of digits to be kept
 could be set for the mean and the standard deviation.
}
\examples{
age = c(18, 19, 22, 19, 26, 23, 18, 20, 22, 19)
strMeanSd(age)
}
\author{
Guillaume T. Vallet \email{gtvallet@gmail.com}, University of Montreal (Canada)
}
\keyword{data,}
\keyword{descriptive,}
\keyword{mean,}
\keyword{std}

