% Generated by roxygen2 (4.0.2): do not edit by hand
\name{descTab}
\alias{descTab}
\title{Compute and Format a data frame into a table suitable for
  publication with mean and standard deviation into parentheses}
\usage{
descTab(data, grp = NULL, toname = NULL, transpose = T)
}
\arguments{
\item{A}{data frame with only the columns to use to compute the
mean and the standard deviation as well as the column to use to
compute by group.}

\item{Optional,}{the name of the column to use to compute by group.}

\item{Optional,}{a vector of names to use for the groups.}

\item{Optional,}{a logical index to transpose or not the data frame.
Default to True.}
}
\value{
Return a data frame by group (if any) with the mean and the
 standard deviation into parentheses.
}
\description{
Compute and Format a data frame into a table suitable for
  publication with mean and standard deviation into parentheses
}
\examples{
# Create data frame faking a group of young and elderly adults.
data = rbind(
     data.frame(subject=1:10,
         age=round(rnorm(10, mean=23, sd=1.5), 0),
         education=round(rnorm(10, mean=15, sd=2.1), 0),
         group='young'),
     data.frame(subject=11:20,
         age=round(rnorm(10, mean=71, sd=2), 0),
         education=round(rnorm(10, mean=14, sd=1.9), 0),
         group='elderly'))
 # Compute the mean and standard deviation of their age and
 #   education. Exclude the subject column to run the function.
descTab(data[,-1], grp='group', toname=c("Young Adults", "Elderly Adults"))
}
\author{
Guillaume T. Vallet \email{gtvallet@gmail.com}, Universit\'{e} de Montr\'{e}al (Canada)
}
\keyword{data,}
\keyword{descriptive}

