% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getDataFiles}
\alias{getDataFiles}
\title{Load multiple csv files into a unique data frame}
\usage{
getDataFiles(folder, targetCol = NULL, separator = ";", toSkip = 0,
  rowToRead = -1, addSubj = NULL)
}
\arguments{
\item{folder}{A string indicating the relative path of the folder
in which are stored the csv files}

\item{targetCol}{A vector of numbers to precise which columns
should be read. Defaults to all columns.}

\item{separator}{A string indicating the symbol use to
differentiate the columns (, ; \\t). Defaults to ";".}

\item{toSkip}{A number to indicate how many lines (rows) should be
ignored before reading the file. Defaults to 0.}

\item{rowToRead}{A number to indicate how many lines (rows) should
be read from the starting reading line. Defaults to all lines.}

\item{addSubj}{Either a vector of 2 numbers to indicate the first
and last character index of the csv file name to extract the
 subject name or a boolean index to TRUE to add a subject number
 equal to the process order of the file. Defaults to NULL to
 indicate that a subject identification is present.}
}
\value{
Return a data frame with all data bind by rows.
}
\description{
A function to load all csv files within a given directory and merge data into a unique data frame.
}
\examples{
getDataFiles()
}
\author{
Guillaume T. Vallet \email{gtvallet@gmail.com}, Universit\'{e} de Montr\'{e}al (Canada)
}
\keyword{csv,}
\keyword{data,}
\keyword{files}

